# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AreaFada OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    SubscriptionTier:
      type: string
      enum: [free, creator, brand, agency, enterprise]
...
    UserProfile:
      type: object
      required: [id, clerkId, email, displayName, tier, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        tier:
          $ref: "#/components/schemas/SubscriptionTier"
        bio:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        country:
          type: string
        avatarUrl:
          type: string
...